pyproject.toml
[project]
name = "dashcam-sign-detector"
version = "0.1.0"
description = "End-to-end dashcam traffic sign detector: YOLOv10 detector + ResNet classifier fine-tuned on GTSRB."
readme = "README.md"
requires-python = ">=3.13,<3.14"
license = { file = "LICENSE" }
authors = [{ name = "Riley Oest" }]

dependencies = [
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "fastai>=2.7.19",
    "numpy>=2.0",
    "pillow>=11.0",
    "scikit-learn>=1.5",
    "scipy>=1.14",
    "matplotlib>=3.9",
    "pandas>=2.2",
    "kaggle>=1.6",
    "ipython>=8.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "ruff>=0.7",
    "jupyter>=1.1",
]

[project.scripts]
dsd-train-classifier = "dashcam_sign_detector.classifier.train:main"
dsd-eval-classifier = "dashcam_sign_detector.classifier.evaluate:main"
dsd-preprocess-classifier = "dashcam_sign_detector.classifier.preprocessing:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dashcam_sign_detector"]

[tool.uv.sources]
torch = [{ index = "pytorch-cu128" }]
torchvision = [{ index = "pytorch-cu128" }]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
